third_party/WebKit/LayoutTests/fast/dom/shadow/tooltips-in-shadow-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test for Case1: Non-Nested
PASS harness.expected is testRunner.tooltipText
Test for Case 2: Nested
PASS harness.expected is testRunner.tooltipText
PASS successfullyParsed is true

TEST COMPLETE
